% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx.balance.R
\name{approx.balance}
\alias{approx.balance}
\title{Compute approximately balancing weights}
\usage{
approx.balance(M, balance.target, zeta = 0.5,
  allow.negative.weights = FALSE, optimizer = c("mosek", "pogs",
  "quadprog"), use.dual = NULL, bound.gamma = FALSE, verbose = FALSE)
}
\arguments{
\item{M}{the feature matrix, see (*)}

\item{balance.target}{the target solution, see (*)}

\item{zeta}{tuning parameter, see (*)}

\item{allow.negative.weights}{are the gammas allowed to be negative?}

\item{optimizer}{Which optimizer to use? Mosek is a commercial solver, but free
academic licenses are available. Needs to be installed separately.
Pogs runs ADMM and may be useful for large problems, and
must be installed separately. Quadprog is the default
R solver.}

\item{use.dual}{whether balancing should be solved in dual form}

\item{bound.gamma}{whether upper bound on gamma should be imposed}

\item{verbose}{whether the optimizer should print progress information}
}
\value{
gamma, the minimizer of (*)
}
\description{
Returns the minimizer of:
  (1 - zeta) ||gamma||^2 + zeta ||M'gamma - balance.target||_infty^2 (*)
}
