% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/baselines.R
\name{elnet.ate}
\alias{elnet.ate}
\title{Elastic net correction for confounding variables}
\usage{
elnet.ate(X, Y, W, target.pop = c(0, 1), alpha = 0.9)
}
\arguments{
\item{X}{the input features}

\item{Y}{the observed responses}

\item{W}{treatment/control assignment, coded as 0/1}

\item{target.pop}{which population should the treatment effect be estimated for?
           (0, 1): average treatment effect for everyone
           0: average treatment effect for controls
           1: average treatment effect for treated}
}
\value{
ATE estimate
}
\description{
Elastic net correction for confounding variables
}

