% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.R
\name{ipw.ate}
\alias{ipw.ate}
\title{Estimate ATE via inverse propensity weighting}
\usage{
ipw.ate(X, Y, W, target.pop = c(0, 1), eps.threshold = 1/20,
  fit.method = c("elnet", "none"), alpha.fit = 0.9,
  prop.method = c("elnet", "randomforest"), alpha.prop = 0.5,
  estimate.se = FALSE)
}
\arguments{
\item{X}{the input features}

\item{Y}{the observed responses}

\item{W}{treatment/control assignment, coded as 0/1}

\item{target.pop}{which population should the treatment effect be estimated for?
(0, 1): average treatment effect for everyone
0: average treatment effect for controls
1: average treatment effect for treated}

\item{eps.threshold}{cap on the estimated propensities}

\item{fit.method}{the method used to fit mu(x, w) = E[Y | X = x, W = w]}

\item{alpha.fit}{tuning paramter for glmnet in the mu model}

\item{prop.method}{the method used to fit e(x) = P[W = 1 | X]}

\item{alpha.prop}{tuning paramter for glmnet in the propsenity model}
}
\value{
ATE estimate
}
\description{
Estimate ATE via inverse propensity weighting
}
