% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/residual.balance.R
\name{residualBalance.mean}
\alias{residualBalance.mean}
\title{Estimate mean outcome at balance.target via residual balancing}
\usage{
residualBalance.mean(XW, YW, balance.target, allow.negative.weights = FALSE,
  use.active.mask = FALSE, zeta, kappa, fit.method = c("elnet", "none"),
  alpha)
}
\arguments{
\item{XW}{the input features for the sub-population of interest}

\item{YW}{the observed responses for the sub-population of interest}

\item{balance.target}{the desired center of the dataset}

\item{allow.negative.weights}{whether negative gammas are allowed for balancing}

\item{use.active.mask}{whether to penalize imbalance more aggressively along active coordinates}

\item{zeta}{tuning parameter for selecting approximately balancing weights}

\item{kappa}{controls strength of active mask penalty, if used}

\item{fit.method}{the method used to fit mu(x) = E[YW | XW = x]}

\item{alpha}{tuning paramter for glmnet}
}
\value{
Estimate for E[YW | XW = balance.target], along with variance estimate
}
\description{
Estimate mean outcome at balance.target via residual balancing
}

