% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{twostep.lasso.ate}
\alias{twostep.lasso.ate}
\title{Refit lasso correction for confounding variables}
\usage{
twostep.lasso.ate(X, Y, W, target.pop = c(0, 1), fit.propensity = TRUE,
  estimate.se = FALSE)
}
\arguments{
\item{X}{the input features}

\item{Y}{the observed responses}

\item{W}{treatment/control assignment, coded as 0/1}

\item{target.pop}{which population should the treatment effect be estimated for?
(0, 1): average treatment effect for everyone
0: average treatment effect for controls
1: average treatment effect for treated}

\item{fit.propensity}{should propensity model be used for variable selection?}
}
\value{
ATE estimate
}
\description{
Refit lasso correction for confounding variables
}
