% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/approx.balance.R
\name{approx.balance}
\alias{approx.balance}
\title{Compute approximately balancing weights}
\usage{
approx.balance(M, balance.target, zeta = 0.5,
  allow.negative.weights = FALSE)
}
\arguments{
\item{M}{the feature matrix, see (*)}

\item{balance.target}{the target solution, see (*)}

\item{zeta}{tuning parameter, see (*)}

\item{allow.negative.weights}{are the gammas allowed to be negative?}
}
\value{
gamma, the minimzer of (*)
}
\description{
Returns the minimizer of:
  (1 - zeta) ||gamma||^2 + zeta ||M'gamma - balance.target||_infty^2 (*)
}

