% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/residual.balance.R
\name{residualBalance.ate}
\alias{residualBalance.ate}
\title{Estimate ATE via approximate residual balancing}
\usage{
residualBalance.ate(X, Y, W, target.pop = c(0, 1),
  allow.negative.weights = FALSE, zeta = 0.5, fit.method = c("elnet",
  "none"), alpha = 0.9, scale.X = TRUE, estimate.se = FALSE)
}
\arguments{
\item{X}{the input features}

\item{Y}{the observed responses}

\item{W}{treatment/control assignment, coded as 0/1}

\item{target.pop}{which population should the treatment effect be estimated for?
(0, 1): average treatment effect for everyone
0: average treatment effect for controls
1: average treatment effect for treated}

\item{allow.negative.weights}{whether negative gammas are allowed for balancing}

\item{zeta}{tuning parameter for selecting approximately balancing weights}

\item{fit.method}{the method used to fit mu(x, w) = E[Y | X = x, W = w]}

\item{alpha}{tuning paramter for glmnet}

\item{scale.X}{whether non-binary features should be noramlized}

\item{estimate.se}{...}
}
\value{
ATE estimate, along with (optional) variance estimate
}
\description{
Estimate ATE via approximate residual balancing
}

