% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.balance.R
\name{residualBalance.mean}
\alias{residualBalance.mean}
\title{Estimate mean outcome at balance.target via residual balancing}
\usage{
residualBalance.mean(XW, YW, balance.target, allow.negative.weights = FALSE,
  zeta, fit.method = c("elnet", "none"), alpha, optimizer = c("pogs",
  "quadprog", "mosek"), use.dual = NULL, verbose = FALSE)
}
\arguments{
\item{XW}{the input features for the sub-population of interest}

\item{YW}{the observed responses for the sub-population of interest}

\item{balance.target}{the desired center of the dataset}

\item{allow.negative.weights}{whether negative gammas are allowed for balancing}

\item{zeta}{tuning parameter for selecting approximately balancing weights}

\item{fit.method}{the method used to fit mu(x) = E[YW | XW = x]}

\item{alpha}{tuning paramter for glmnet}

\item{optimizer}{which optimizer to use for approximate balancing}

\item{use.dual}{whether balancing should be solved in dual form}

\item{verbose}{whether the optimizer should print progress information}
}
\value{
Estimate for E[YW | XW = balance.target], along with variance estimate
}
\description{
Estimate mean outcome at balance.target via residual balancing
}
